+++ breeze-dsl/src/main/kotlin/com/windea/breezeframework/dsl/markup/JsonDsl.kt
@@ -11,6 +11,7 @@ import com.windea.breezeframework.dsl.markup.JsonConfig.quote
 
 //REGION top annotations and interfaces
 
+/**Json的Dsl。*/
 @DslMarker
 private annotation class JsonDsl